requirements.txt
black==23.1.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
idna==3.4
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.0
platformdirs==3.0.0
python-dotenv==1.0.0
requests==2.28.2
tomli==2.0.1
urllib3==1.26.14
websocket-client==1.5.1
